Set up Localvimrc plugin
This enables me to use some custom `.vimrc` configs to match changes we
made to our linters at work, without having to change my syntastic
linters for all other projects.

https://github.com/embear/vim-localvimrc

"This plugin searches for local vimrc files in the file system
tree of the currently opened file. It searches for all ".lvimrc" files
from the directory of the file up to the root directory. By default
those files are loaded in order from the root directory to the directory
of the file. The filename and amount of loaded files are customizable
through global variables.

For security reasons it the plugin asks for confirmation before loading
a local vimrc file and loads it using |:sandbox| command. The plugin
asks once per session and local vimrc before loading it, if the file
didn't change since previous loading.

It is possible to define a whitelist and a blacklist of local vimrc
files that are loaded or ignored unconditionally."