Make syntastic a true git submodule
There was something weird about the installed version of `syntastic` in
my bundle directory; it was a directory, and not really a submodule. I
ran into an issue when trying to remove 'ack-vim' which was caused by
the library not being treated as a submodule.

In any case, now it should be easier to manage this plugin.